a21s-common: Remove unecessary powerhint entries
Our device is relatively new, shipping in 2020, and so
our hardware can mostly support Android's UI without
much UI stuttering at all. For the sake of battery life,
we can remove most of the boosting entries as it doesn't
provide a noticeable difference at this stage of the device's
lifespan and it will be massively draining the battery.

Change-Id: Ieec6ea3c9e8515f004775aa7abcdda35f1264e01